+++ src/compiler/compile.ts
@@ -1,6 +1,6 @@
 import type { GoslingSpec, TemplateTrackDef, VisUnitApiData } from '@gosling-lang/gosling-schema';
 import type { HiGlassSpec } from '@gosling-lang/higlass-schema';
-import { traverseToFixSpecDownstream, overrideDataTemplates } from './spec-preprocess';
+import { traverseToFixSpecDownstream } from './spec-preprocess';
 import { replaceTrackTemplates } from '../core/utils/template';
 import { getRelativeTrackInfo, type Size } from './bounding-box';
 import type { CompleteThemeDeep } from '../core/utils/theme';
@@ -8,7 +8,6 @@ import type { UrlToFetchOptions } from 'src/core/gosling-component';
 import { renderHiGlass as createHiGlassModels } from './create-higlass-models';
 import { manageResponsiveSpecs } from './responsive';
 import type { IdTable } from '../api/track-and-view-ids';
-import { publish } from '../api/pubsub';
 
 /** The callback function called everytime after the spec has been compiled */
 export type CompileCallback = (
@@ -32,9 +31,6 @@ export function compile(
 ) {
     // Make sure to keep the original spec as-is
     const specCopy = JSON.parse(JSON.stringify(spec));
-
-    // Override default visual encoding (i.e., `DataTrack` => `BasicSingleTrack`)
-    overrideDataTemplates(specCopy);
 
     // Replace track templates with raw gosling specs (i.e., `TemplateTrack` => `SingleTrack | OverlaidTrack`)
     replaceTrackTemplates(specCopy, templates);
@@ -69,13 +65,6 @@ export function compile(
         traverseToFixSpecDownstream(specCopy);
         trackInfos = getRelativeTrackInfo(specCopy, theme).trackInfos;
     }
-
-    // publish the fixed spec
-    // used for alt-gosling
-    publish('specProcessed', {
-        id: specCopy.id,
-        spec: specCopy
-    });
 
     // Make HiGlass models for individual tracks
     createHiGlassModels(specCopy, trackInfos, callback, theme, urlToFetchOptions);